Added functions to import missing telemetry.
When you switch to Self-Powered mode both the mobile app and API block the data for time of use, therefore my normal scripts don't import some of the telemetry.  However it's still logged and you can retrieve it when you switch back to Time-Based control.  I added the ability to import it after situations when I need to be in Self-Powered mode.