More changes for access token updates.
Not only has Tesla changed the format of the access token, the old method of retieving a new access token from the refresh token has also changed.  Also, the access token no longer lasts for 45 days.  The new access token expires in 8 hours.  The refresh token expiration is unknown and seems to stay constant even after getting a new access token which is not how it worked previously.  Previously you would get a new refresh token with a new access token.  These changes now check, on an hourly crontab, the expiration date and will refresh the access token and write it to a new encrypted token config file 2 hours before the expiration.