Code refactoring
Updated to retrieve the entire range of values in the Google Sheet and avoid repetitive Google API calls for each cell value in the same function.  Deprecated a couple helper functions to make those repetitive Google API calls.  Also, reused a common function to schedule an event for the following morning, which also now accounts for Daylight Savings.