Code refactoring
Consolidated all the logic for calculating scheduled departure time for 2 vehicles based on location into a function.  Previously this was partially done inside a Google Sheet which was a poor design.  The only things left in the Google Sheet are the configurable items:  Charging finish time and email notification on/off.  The charge rates can be added in the Google Sheet in the future if these need to be changed as configurations.