Changed operating mode logic.
Leaving the mode the same (Time-Based Control) and just changing the backup reserve (100%) to achieve the same behavior as Back-up Only mode, without changing to that mode.  There is an intermittent issue where if you change the system to Back-up Only mode and you try to get that day's data from the API, it will return an empty response.